refactor: extract logic shared by adlibs(actions) handlers
