+++ packages/live-status-gateway/src/collections/globalAdLibActionsHandler.ts
@@ -1,28 +1,10 @@
 import { Logger } from 'winston'
 import { CoreHandler } from '../coreHandler'
-import { Collection } from '../wsHandler'
-import { PublicationCollection } from '../publicationCollection'
-import { RundownBaselineAdLibAction } from '@sofie-automation/corelib/dist/dataModel/RundownBaselineAdLibAction'
 import { CollectionName } from '@sofie-automation/corelib/dist/dataModel/Collections'
 import { CorelibPubSub } from '@sofie-automation/corelib/dist/pubsub'
-import { DBRundownPlaylist } from '@sofie-automation/corelib/dist/dataModel/RundownPlaylist'
-import { CollectionHandlers } from '../liveStatusServer'
-import { RundownId } from '@sofie-automation/corelib/dist/dataModel/Ids'
-import { PickKeys } from '@sofie-automation/shared-lib/dist/lib/types'
+import { RundownContentHandlerBase } from './rundownContentHandlerBase'
 
-const PLAYLIST_KEYS = ['currentPartInfo', 'nextPartInfo'] as const
-type Playlist = PickKeys<DBRundownPlaylist, typeof PLAYLIST_KEYS>
-
-export class GlobalAdLibActionsHandler
-	extends PublicationCollection<
-		RundownBaselineAdLibAction[],
-		CorelibPubSub.rundownBaselineAdLibActions,
-		CollectionName.RundownBaselineAdLibActions
-	>
-	implements Collection<RundownBaselineAdLibAction[]>
-{
-	private _currentRundownId: RundownId | undefined
-
+export class GlobalAdLibActionsHandler extends RundownContentHandlerBase<CorelibPubSub.rundownBaselineAdLibActions> {
 	constructor(logger: Logger, coreHandler: CoreHandler) {
 		super(
 			CollectionName.RundownBaselineAdLibActions,
@@ -30,36 +12,5 @@ export class GlobalAdLibActionsHandler
 			logger,
 			coreHandler
 		)
-	}
-
-	init(handlers: CollectionHandlers): void {
-		super.init(handlers)
-
-		handlers.playlistHandler.subscribe(this.onPlaylistUpdate, PLAYLIST_KEYS)
-	}
-
-	protected changed(): void {
-		this.updateAndNotify()
-	}
-
-	private onPlaylistUpdate = (data: Playlist | undefined): void => {
-		this.logUpdateReceived('playlist')
-		const prevRundownId = this._currentRundownId
-		const rundownPlaylist = data
-
-		this._currentRundownId = rundownPlaylist?.currentPartInfo?.rundownId ?? rundownPlaylist?.nextPartInfo?.rundownId
-
-		if (prevRundownId !== this._currentRundownId) {
-			this.stopSubscription()
-			if (this._currentRundownId) {
-				this.setupSubscription([this._currentRundownId])
-			}
-		}
-	}
-
-	private updateAndNotify(): void {
-		const collection = this.getCollectionOrFail()
-		this._collectionData = collection.find({ rundownId: this._currentRundownId })
-		this.notify(this._collectionData)
 	}
 }
